Changed name to comparative_genetics.py
